+++ packages/polkadex-api/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./base-api";
+export * from "./types";
+export * from "./api";
 export * from "./modules";