+++ packages/ui/src/components/index.ts
@@ -34,3 +34,7 @@ export * from "./toggleGroup";
 export * from "./filterGroup";
 export * from "./scrollArea";
 export * from "./resizable";
+export * from "./calendar";
+export * from "./datePicker";
+export * from "./carousel";
+export * from "./interactable";